Refactor: Organize TV show episode remote response models
