+++ data/src/main/java/com/london/data/datasource/remote/details/tvshowdetails/TvShowDetailsRemoteDataSource.kt
@@ -2,7 +2,7 @@ package com.london.data.datasource.remote.details.tvshowdetails
 
 import com.london.data.datasource.remote.details.tvshowdetails.model.TvShowCastRemoteResponse
 import com.london.data.datasource.remote.details.tvshowdetails.model.TvShowDetailsRemoteResponse
-import com.london.data.datasource.remote.details.tvshowdetails.model.TvShowEpisodesRemoteResponse
+import com.london.data.datasource.remote.details.tvshowdetails.model.tvshowepisode.TvShowEpisodesRemoteResponse
 import com.london.data.datasource.remote.details.tvshowdetails.model.TvShowImagesRemoteResponse
 
 interface TvShowDetailsRemoteDataSource {
