+++ data/src/main/java/com/london/data/datasource/remote/details/tvshowdetails/model/tvshowepisode/EpisodeGuestStar.kt
@@ -1,0 +1,30 @@
+package com.london.data.datasource.remote.details.tvshowdetails.model.tvshowepisode
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class EpisodeGuestStar(
+    @SerialName("character")
+    val character: String,
+    @SerialName("credit_id")
+    val creditId: String,
+    @SerialName("order")
+    val order: Int,
+    @SerialName("adult")
+    val adult: Boolean,
+    @SerialName("gender")
+    val gender: Int,
+    @SerialName("id")
+    val id: Int,
+    @SerialName("known_for_department")
+    val knownForDepartment: String,
+    @SerialName("name")
+    val name: String,
+    @SerialName("original_name")
+    val originalName: String,
+    @SerialName("popularity")
+    val popularity: Double,
+    @SerialName("profile_path")
+    val profilePath: String?
+)
